Fixed scoring for other languages.
French, German, and probably jp all use a different numbering system. And since it's stored as a string *-sighs-*

Went through and added a check to all of these to return a proper value on other languages.

Also added a check for critical missions specifically (they return a 0/1 vs an actual score)