Added Collectable Shop + Base Shop Info
Added info to grab from the collectable shop (InclusionShop) and the basic shops (ShopExchangeCurrency).

There's... some minor wonkyness that comes with the basic shop. The base addon info is different than the collectables (aka all the data is scattered around in the ATK values) and there isn't a good way to get the currencyId of what is required. I'll go through and add the itemExchange versions of it later because those also read differently than the basic gil/moon currency shops.