Update Callback.cs
Replace signature hooking with FireCallback delegate exposed in CS.
Kept FireRaw method for compatibility since it's public, but could be marked as obsolete.
DecodeValues could have argument cnt changed to uint but I left it as is for compatibility. Technically not an issue.
Might also want to use Convert.ToUInt32() instead of implicit conversions from int to uint, but it would mean method could theorically throw, so implicit cast used for compatibility.